refactor: re-export the login surface from the auth package
auth/__init__ now carries AuthMethod/PasswordMethod/safe_next_path (a
Macro-5a addition the package never re-exported), and the five imports
that reached into .auth.auth_method now ride the package __init__,
matching the uniform pattern every sibling package follows.
