+++ src/genro_asgi_core/auth/__init__.py
@@ -12,16 +12,20 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Auth capability package: the credential core and the server-side mixin.
+"""Auth capability package: the credential core, the mixin, and the login surface.
 
 ``AuthCore`` (basic/bearer/jwt verification) and ``AuthMixin`` (the §5.5
 identity precedence over sessions). ``AuthMiddleware`` — the chain entry point
-armed by the mixin — lives in ``middleware/authentication.py``.
+armed by the mixin — lives in ``middleware/authentication.py``. The login
+surface — the self-describing ``AuthMethod``/``PasswordMethod`` (core 1d) and
+the ``safe_next_path`` open-redirect guard — is re-exported here too, so every
+consumer imports auth symbols from the package, never from its modules.
 """
 
 from __future__ import annotations
 
 from .api_key_store import ApiKeyStore, FileApiKeyStore
+from .auth_method import AuthMethod, PasswordMethod, safe_next_path
 from .core import AuthCore
 from .mixin import AuthMixin
 from .user_store import FileUserStore, UserStore
@@ -29,8 +33,11 @@ from .user_store import FileUserStore, UserStore
 __all__ = [
     "ApiKeyStore",
     "AuthCore",
+    "AuthMethod",
     "AuthMixin",
     "FileApiKeyStore",
     "FileUserStore",
+    "PasswordMethod",
     "UserStore",
+    "safe_next_path",
 ]
